requirements.txt
aiohttp==3.6.2
aiosqlite==0.16.1
aiosqlite3==0.3.0
astroid==2.4.2
async-timeout==3.0.1
attrs==20.2.0
certifi==2020.12.5
chardet==3.0.4
colorama==0.4.3
dblpy==0.4.0
discord-interactive==4.0
discord.py==1.6.0
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
multidict==4.7.6
Pillow==8.0.1
psutil==5.8.0
pylint==2.6.0
ratelimiter==1.2.0.post0
requests==2.25.1
six==1.15.0
toml==0.10.1
typing-extensions==3.7.4.3
urllib3==1.26.2
wrapt==1.12.1
yarl==1.5.1
